fixed file names and deleted obsolete code
